pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.4.0",
    "streamlit>=1.42.2",
    "twilio>=9.4.6",
]
